[CoordinatedGraphics] Remove CoordinatedGraphicsScene https://bugs.webkit.org/show_bug.cgi?id=286677
Reviewed by Alejandro G. Castro.

At this moment CoordinatedGraphicsScene is only used to paint layers
into the texture mapper. It could be just added to the
ThreadedCompositor class.

* Source/WebCore/platform/graphics/texmap/TextureMapperAnimation.cpp:
(WebCore::TextureMapperAnimation::apply):
* Source/WebKit/PlatformPlayStation.cmake:
* Source/WebKit/SourcesGTK.txt:
* Source/WebKit/SourcesWPE.txt:
* Source/WebKit/WebProcess/WebPage/CoordinatedGraphics/CoordinatedGraphicsScene.cpp: Removed.
* Source/WebKit/WebProcess/WebPage/CoordinatedGraphics/CoordinatedGraphicsScene.h: Removed.
* Source/WebKit/WebProcess/WebPage/CoordinatedGraphics/LayerTreeHost.cpp:
(WebKit::LayerTreeHost::requestComposition):
* Source/WebKit/WebProcess/WebPage/CoordinatedGraphics/ThreadedCompositor.cpp:
(WebKit::ThreadedCompositor::invalidate):
(WebKit::ThreadedCompositor::setDamagePropagation):
(WebKit::ThreadedCompositor::updateSceneState):
(WebKit::ThreadedCompositor::paintToCurrentGLContext):
(WebKit::ThreadedCompositor::renderLayerTree):
(WebKit::ThreadedCompositor::requestComposition):
(WebKit::ThreadedCompositor::scheduleUpdate):
(WebKit::ThreadedCompositor::updateViewport): Deleted.
(WebKit::ThreadedCompositor::addSurfaceDamage): Deleted.
(WebKit::ThreadedCompositor::updateScene): Deleted.
* Source/WebKit/WebProcess/WebPage/CoordinatedGraphics/ThreadedCompositor.h:

Canonical link: https://commits.webkit.org/289555@main
