+++ Source/WebKit/PlatformPlayStation.cmake
@@ -150,7 +150,6 @@ endif ()
 if (USE_COORDINATED_GRAPHICS)
     list(APPEND WebKit_SOURCES
         WebProcess/WebPage/CoordinatedGraphics/CompositingRunLoop.cpp
-        WebProcess/WebPage/CoordinatedGraphics/CoordinatedGraphicsScene.cpp
         WebProcess/WebPage/CoordinatedGraphics/CoordinatedSceneState.cpp
         WebProcess/WebPage/CoordinatedGraphics/LayerTreeHost.cpp
         WebProcess/WebPage/CoordinatedGraphics/ThreadedCompositor.cpp
