+++ Source/WebKit/WebProcess/WebPage/CoordinatedGraphics/LayerTreeHost.cpp
@@ -412,7 +412,7 @@ void LayerTreeHost::requestComposition()
     }
 #endif
 
-    m_compositor->updateScene();
+    m_compositor->scheduleUpdate();
 }
 
 RunLoop* LayerTreeHost::compositingRunLoop() const
