+++ Source/WebKit/WebProcess/WebPage/CoordinatedGraphics/ThreadedCompositor.cpp
@@ -33,7 +33,10 @@
 #include "LayerTreeHost.h"
 #include "WebPage.h"
 #include "WebProcess.h"
+#include <WebCore/CoordinatedPlatformLayer.h>
+#include <WebCore/Damage.h>
 #include <WebCore/PlatformDisplay.h>
+#include <WebCore/TextureMapperLayer.h>
 #include <WebCore/TransformationMatrix.h>
 #include <wtf/SetForScope.h>
 #include <wtf/SystemTracing.h>
@@ -81,6 +84,7 @@ ThreadedCompositor::ThreadedCompositor(LayerTreeHost& layerTreeHost, ThreadedDis
 #endif
     : m_layerTreeHost(&layerTreeHost)
     , m_surface(AcceleratedSurface::create(*this, layerTreeHost.webPage(), [this] { frameComplete(); }))
+    , m_sceneState(&m_layerTreeHost->sceneState())
     , m_flipY(m_surface->shouldPaintMirrored())
     , m_compositingRunLoop(makeUnique<CompositingRunLoop>([this] { renderLayerTree(); }))
 #if HAVE(DISPLAY_LINK)
@@ -98,7 +102,7 @@ ThreadedCompositor::ThreadedCompositor(LayerTreeHost& layerTreeHost, ThreadedDis
     m_display.displayUpdate = { 0, c_defaultRefreshRate / 1000 };
 #endif
 
-    m_compositingRunLoop->performTaskSync([this, protectedThis = Ref { *this }, sceneState = Ref { m_layerTreeHost->sceneState() }] {
+    m_compositingRunLoop->performTaskSync([this, protectedThis = Ref { *this }] {
 #if !HAVE(DISPLAY_LINK)
         m_display.updateTimer = makeUnique<RunLoop::Timer>(RunLoop::current(), this, &ThreadedCompositor::displayUpdateFired);
 #if USE(GLIB_EVENT_LOOP)
@@ -107,8 +111,6 @@ ThreadedCompositor::ThreadedCompositor(LayerTreeHost& layerTreeHost, ThreadedDis
 #endif
         m_display.updateTimer->startOneShot(Seconds { 1.0 / m_display.displayUpdate.updatesPerSecond });
 #endif
-
-        m_scene = adoptRef(new CoordinatedGraphicsScene(*this, sceneState.get()));
 
         // GLNativeWindowType depends on the EGL implementation: reinterpret_cast works
         // for pointers (only if they are 64-bit wide and not for other cases), and static_cast for
@@ -137,7 +139,6 @@ uint64_t ThreadedCompositor::surfaceID() const
 void ThreadedCompositor::invalidate()
 {
     ASSERT(RunLoop::isMain());
-    m_scene->detach();
     m_compositingRunLoop->stopUpdates();
 #if HAVE(DISPLAY_LINK)
     m_didRenderFrameTimer.stop();
@@ -148,20 +149,20 @@ void ThreadedCompositor::invalidate()
         if (!m_context || !m_context->makeContextCurrent())
             return;
 
-        // Update the scene at this point ensures the layers state are correctly propagated
-        // in the ThreadedCompositor and in the CompositingCoordinator.
-        m_scene->updateSceneState();
+        // Update the scene at this point ensures the layers state are correctly propagated.
+        updateSceneState();
 
-        m_scene->purgeGLResources();
+        m_sceneState->invalidateCommittedLayers();
+        m_textureMapper = nullptr;
         m_surface->willDestroyGLContext();
         m_context = nullptr;
         m_surface->finalize();
-        m_scene = nullptr;
 
 #if !HAVE(DISPLAY_LINK)
         m_display.updateTimer = nullptr;
 #endif
     });
+    m_sceneState = nullptr;
     m_layerTreeHost = nullptr;
     m_surface->willDestroyCompositingRunLoop();
     m_compositingRunLoop = nullptr;
@@ -209,26 +210,80 @@ void ThreadedCompositor::preferredBufferFormatsDidChange()
 }
 #endif
 
-void ThreadedCompositor::updateViewport()
-{
-    m_compositingRunLoop->scheduleUpdate();
-}
-
 #if ENABLE(DAMAGE_TRACKING)
-void ThreadedCompositor::setDamagePropagation(WebCore::Damage::Propagation damagePropagation)
+void ThreadedCompositor::setDamagePropagation(Damage::Propagation damagePropagation)
 {
-    m_scene->setDamagePropagation(damagePropagation);
-}
-
-const Damage& ThreadedCompositor::addSurfaceDamage(const Damage& damage)
-{
-    return m_surface->addDamage(damage);
+    m_damagePropagation = damagePropagation;
 }
 #endif
 
+void ThreadedCompositor::updateSceneState()
+{
+    if (!m_textureMapper)
+        m_textureMapper = TextureMapper::create();
+
+    m_sceneState->rootLayer().flushCompositingState(*m_textureMapper);
+    for (auto& layer : m_sceneState->committedLayers())
+        layer->flushCompositingState(*m_textureMapper);
+}
+
+void ThreadedCompositor::paintToCurrentGLContext(const TransformationMatrix& matrix, const IntSize& size)
+{
+    updateSceneState();
+
+    FloatRect clipRect(FloatPoint { }, size);
+    TextureMapperLayer& currentRootLayer = m_sceneState->rootLayer().ensureTarget();
+    if (currentRootLayer.transform() != matrix)
+        currentRootLayer.setTransform(matrix);
+
+    bool sceneHasRunningAnimations = currentRootLayer.applyAnimationsRecursively(MonotonicTime::now());
+
+    m_textureMapper->beginPainting(m_flipY ? TextureMapper::FlipY::Yes : TextureMapper::FlipY::No);
+    m_textureMapper->beginClip(TransformationMatrix(), FloatRoundedRect(clipRect));
+
+    std::optional<FloatRoundedRect> rectContainingRegionThatActuallyChanged;
+#if ENABLE(DAMAGE_TRACKING)
+    currentRootLayer.prepareForPainting(*m_textureMapper);
+    if (m_damagePropagation != Damage::Propagation::None) {
+        Damage frameDamage;
+        WTFBeginSignpost(this, CollectDamage);
+        currentRootLayer.collectDamage(*m_textureMapper, frameDamage);
+        WTFEndSignpost(this, CollectDamage);
+
+        if (m_damagePropagation == Damage::Propagation::Unified) {
+            Damage boundsDamage;
+            boundsDamage.add(frameDamage.bounds());
+            frameDamage = WTFMove(boundsDamage);
+        }
+
+        const auto& damageSinceLastSurfaceUse = m_surface->addDamage(!frameDamage.isInvalid() && !frameDamage.isEmpty() ? frameDamage : Damage::invalid());
+        if (!damageSinceLastSurfaceUse.isInvalid() && !FloatRect(damageSinceLastSurfaceUse.bounds()).contains(clipRect))
+            rectContainingRegionThatActuallyChanged = FloatRoundedRect(damageSinceLastSurfaceUse.bounds());
+    }
+#endif
+
+    if (rectContainingRegionThatActuallyChanged)
+        m_textureMapper->beginClip(TransformationMatrix(), *rectContainingRegionThatActuallyChanged);
+
+    WTFBeginSignpost(this, PaintTextureMapperLayerTree);
+    currentRootLayer.paint(*m_textureMapper);
+    WTFEndSignpost(this, PaintTextureMapperLayerTree);
+
+    if (rectContainingRegionThatActuallyChanged)
+        m_textureMapper->endClip();
+
+    m_fpsCounter.updateFPSAndDisplay(*m_textureMapper, clipRect.location(), matrix);
+
+    m_textureMapper->endClip();
+    m_textureMapper->endPainting();
+
+    if (sceneHasRunningAnimations)
+        scheduleUpdate();
+}
+
 void ThreadedCompositor::renderLayerTree()
 {
-    ASSERT(m_scene);
+    ASSERT(m_sceneState);
     ASSERT(m_compositingRunLoop->isCurrent());
 #if PLATFORM(GTK) || PLATFORM(WPE)
     TraceScope traceScope(RenderLayerTreeStart, RenderLayerTreeEnd);
@@ -245,7 +300,7 @@ void ThreadedCompositor::renderLayerTree()
 #endif
 
     // Retrieve the scene attributes in a thread-safe manner.
-    WebCore::IntSize viewportSize;
+    IntSize viewportSize;
     float deviceScaleFactor;
     uint32_t compositionRequestID;
     {
@@ -256,7 +311,7 @@ void ThreadedCompositor::renderLayerTree()
 
 #if !HAVE(DISPLAY_LINK)
         // Client has to be notified upon finishing this scene update.
-        m_attributes.clientRendersNextFrame = m_scene->state().layersDidChange();
+        m_attributes.clientRendersNextFrame = m_sceneState->layersDidChange();
 #endif
     }
 
@@ -284,7 +339,7 @@ void ThreadedCompositor::renderLayerTree()
     m_surface->clearIfNeeded();
 
     WTFBeginSignpost(this, PaintToGLContext);
-    m_scene->paintToCurrentGLContext(viewportTransform, FloatRect { FloatPoint { }, viewportSize }, m_flipY);
+    paintToCurrentGLContext(viewportTransform, viewportSize);
     WTFEndSignpost(this, PaintToGLContext);
 
     WTFEmitSignpost(this, DidRenderFrame, "compositionResponseID %i", compositionRequestID);
@@ -318,11 +373,11 @@ uint32_t ThreadedCompositor::requestComposition()
         m_attributes.viewportSize.scale(m_attributes.deviceScaleFactor);
         compositionRequestID = ++m_attributes.compositionRequestID;
     }
-    m_compositingRunLoop->scheduleUpdate();
+    scheduleUpdate();
     return compositionRequestID;
 }
 
-void ThreadedCompositor::updateScene()
+void ThreadedCompositor::scheduleUpdate()
 {
     m_compositingRunLoop->scheduleUpdate();
 }
